requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.4
Django==4.2.20
djangorestframework==3.15.2
djlint==1.36.4
EditorConfig==0.17.0
idna==3.10
jsbeautifier==1.15.4
json5==0.10.0
pathspec==0.12.1
psycopg2-binary==2.9.10
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
